+++ esp_cam_sensor/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added option to return byte swap flags for GC0308
+- Added luma para in ov2710 driver.
 
 ## 1.4.0